refactor: move to model package
